move `sample_unique` method from `Panel` to `Component`
The `Panel` version of this method did nothing panel specific
except for accessing the panel given by the index `c`.  This
did not seem like enough dependency on the structure of `Panel`
to justify not putting the method on `Component`.  A wrapper
could be introduced that does only this work and delegates to
the `Component` method, but it is used in only one location.