remove out-of-date comments; move `initial_constraints` attribute to the Matrix class
The `initial_constraint` attribute is saved from the settings of `base` (the
argument to Matrix.__init__) and is untouched until alternatives to the
answer panel are being generated.  It was stored on the component.  This
means it was stored nine times when it really only needed to be stored
once.  A similar story is probably true about the uniformity attribute
(it is sampled once per puzzle), but it is more embeded in methods
across modules.