update README, string formatting of `Matrix`/`Rules`; expose `n_alternatives` parameter
I wrote a readme that explains in more detail the purpose
of this repository and its expected usage, as well as providing
a disclaimer about scientific replication efforts that attempt to
utilize this code.  In the process, I realized that I might as well
expose the alternatives generation method and not include it as
part of RPM initialization so that end users can choose the number
of wrong answers they would like to generate.  In addition, I realized
that the string formatting of `Matrix` and `Rules` did not log quite
as much information as might be needed to understand the structure
of the puzzle from the output, alone.  I amended this so that these
strings can serve as complete, human-readable reports.