support rendering of RPMs with user-supplied background colors; permit user-supplied rulesets
When RPMs are used in image generation contexts, it can be useful to
alter or randomize the background color of the grid to reduce the
incentive to reproduce any given color.  This is an uninvasive change
as the background color plays no role in determining the correctness
of a grid.

Though the parameter was exposed, user-provided custom rulesets were
being accidentally ignored.  This was also corrected.