demonstrate use of `background_color` parameter; correct `background_color`'s color interpretation
Internally, most colors are treated as `255 - color_value` rather than
as providing `color_value`.  This was accidentally applied to the
`background_color` parameter as well.  This was corrected so that the
background color is rendered as provided and the usage of this
(non-optional) parameter was added to the short interpreter example
in the README.