save images of exactly the provided size
Previously I thought it was fine if, when the specified `image_size`
provided to `Matrix.save` was not divisible by 3, I rounded down to
the next biggest image size that is.  This was inappropriate in a
downstream application, so I am correcting it.  The excess background
space that cannot be evenly spread across panels (i.e. the bottom
1 or 2 pixels) are now allowed to sit below the last panel.  All line
drawing (polygon borders not included) has been moved into `generate_matrix`
from `render_panel` so that it may be applied to these excess pixels rather
than stopping short where the bottom row's panel overlays end.